backend/requirement.txt
fastapi
uvicorn
python-dotenv
python-multipart
jose
torchaudio
torch
pydantic
celery
redis